+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- DoMINO model architecture, datapipe and training recipe
 - DrivAerML dataset support in FIGConvNet example.
 
 ### Changed